include band centers in main code
