+++ src/NoisySignalIntegration.jl
@@ -20,6 +20,7 @@ include("curves.jl")
 include("bounds.jl")
 include("noise.jl")
 include("integration.jl")
+include("bandcenter.jl")
 include("plotting.jl")
 include("testdata.jl")
 
@@ -39,6 +40,7 @@ export
     stitch,
     fit_noise,
     get_cov,
+    mc_bandcenter,
     mc_integrate,
     mean,
     plotautocovfit,
